discord_bots/requirements.txt
black
discord.py
pyyaml
openai